# Leaf polygons as sorted vector