This PR improve performance on MinGW builds.