@lmichaelis 2023-07-23
Thanks for the PR @Try! I still can reproduce any performance improvement but I don't see a degradation either so either way I'm happy to merge. Here are my benchmark results:

```
:: CLANG

OLD (RELEASE) = 0.206s
NEW (RELEASE) = 0.207s

OLD (DEBUG) = 1.510s
NEW (DEBUG) = 1.759s

:: GCC

OLD (RELEASE) = 0.207s
NEW (RELEASE) = 0.210s

OLD (DEBUG) = 1.503s
NEW (DEBUG) = 1.583s
```